requirements.txt
streamlit
tensorflow>=2.0.0
numpy
Pillow
